Height

+++ 9-binary_tree_height.c
@@ -9,8 +9,8 @@
 
 size_t binary_tree_height(const binary_tree_t *tree)
 {
-	int right_knot = 0;
-	int left_knot = 0;
+	size_t right_knot = 0;
+	size_t left_knot = 0;
 
 	if (tree != NULL)
 	{
